apps/mcp-servers/google-search-console/pyproject.toml
[project]
name = "google-search-console-mcp"
version = "0.1.0"
description = "MCP server for Google Search Console — search analytics, URL inspection, and sitemap management"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.2.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
